Time: 2 ms (98.70%) | Memory: 43.8 MB (51.71%) - LeetSync

+++ 3471-minimum-average-of-smallest-and-largest-elements/minimum-average-of-smallest-and-largest-elements.java
@@ -9,6 +9,6 @@ class Solution {
        
        m1=Math.min(m1,m);
     }
-    return (double)m1;
+    return m1;
     }
 }
